refactor(storage): extract SlowLogRepository across all three storage adapters
